refactor: move sos styles to css
