+++ src/components/Sos.css
@@ -1,0 +1,19 @@
+.sos-container {
+  min-height: 70vh;
+  display: grid;
+  place-items: center;
+  padding: 16px;
+}
+
+.sos-button {
+  width: 18rem;
+  height: 18rem;
+  border-radius: 9999px;
+  background: #dc2626;
+  color: #fff;
+  font-weight: 800;
+  font-size: 56px;
+  box-shadow: 0 20px 50px -12px rgba(220, 38, 38, 0.6);
+  border: 8px solid rgba(248, 113, 113, 0.5);
+  cursor: pointer;
+}
